requirements.txt
pandas==1.5.2
numpy==1.24.1
matplotlib==3.6.2
seaborn==0.11.2
scikit-learn==1.0.2
scipy==1.8.0
requests==2.28.1
time==3.9.0
json==2.0.9
telethon==1.24.0
python-dotenv==0.19.2